+++ src/test/java/org/amoseman/budgetingbackend/service/FinanceRecordServiceTest.java
@@ -6,6 +6,7 @@ import org.amoseman.budgetingbackend.dao.impl.sql.FinanceRecordDAOImpl;
 import org.amoseman.budgetingbackend.database.DatabaseConnection;
 import org.amoseman.budgetingbackend.database.impl.sql.sqlite.DatabaseConnectionImpl;
 import org.amoseman.budgetingbackend.exception.FinanceRecordAlreadyExistsException;
+import org.amoseman.budgetingbackend.exception.FinanceRecordDoesNotExistException;
 import org.amoseman.budgetingbackend.exception.NegativeValueException;
 import org.amoseman.budgetingbackend.pojo.record.Expense;
 import org.amoseman.budgetingbackend.pojo.record.Income;
@@ -65,6 +66,27 @@ class FinanceRecordServiceTest {
         assertEquals("category", income.category);
         assertEquals("description", income.description);
 
+        try {
+            financeRecordService.updateIncome("alice", income.uuid, new IncomeInfo(0, 2024, 1, 1, "", ""));
+        }
+        catch (NegativeValueException | FinanceRecordDoesNotExistException e) {
+            fail(e);
+        }
+        records = financeRecordService.getIncome(
+                "alice",
+                "2023",
+                "1",
+                "1",
+                "2024",
+                "1",
+                "2"
+        );
+        if (records.isEmpty()) {
+            fail("Unable to retrieve income after update");
+        }
+        Income updated = records.get(0);
+        assertEquals(0, updated.amount);
+
         boolean deleted = financeRecordService.removeIncome("alice", income.uuid);
         if (!deleted) {
             fail("Unable to deleted added income");
@@ -117,6 +139,27 @@ class FinanceRecordServiceTest {
         assertEquals("category", expense.category);
         assertEquals("description", expense.description);
         assertEquals("bucket", expense.bucket);
+
+        try {
+            financeRecordService.updateExpense("alice", expense.uuid, new ExpenseInfo(0, 2024, 1, 1, "", "", ""));
+        }
+        catch (NegativeValueException | FinanceRecordDoesNotExistException e) {
+            fail(e);
+        }
+        records = financeRecordService.getExpenses(
+                "alice",
+                "2023",
+                "1",
+                "1",
+                "2024",
+                "1",
+                "2"
+        );
+        if (records.isEmpty()) {
+            fail("Unable to retrieve expense after update");
+        }
+        Expense updated = records.get(0);
+        assertEquals(0, updated.amount);
 
         boolean deleted = financeRecordService.removeExpense("alice", expense.uuid);
         if (!deleted) {
